code-review fixes + react router
