+++ README.md
@@ -1,1 +1,5 @@
 # Imaginary Realtor Agency
+
+npm start - start project
+npm build - create a build for deployment
+npm test - test project
