+++ src/styled/GlobalStyles.js
@@ -1,0 +1,1 @@
+import styled from 'styled-components';
