+++ src/utils.js
@@ -1,0 +1,3 @@
+export const priceHandler = (price) => {
+  return `$${price.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',')}`;
+};
